refactor: ClothesStoreDatabase 생성
- CartDatabase 제거
- ClothesStoreDatabase 추가
