--- app/src/main/java/woowacourse/shopping/data/cart/CartDatabase.kt
@@ -1,29 +1,0 @@
-package woowacourse.shopping.data.cart
-
-import android.content.Context
-import androidx.room.Database
-import androidx.room.Room
-import androidx.room.RoomDatabase
-import woowacourse.shopping.data.product.ProductEntity
-
-@Database(entities = [ProductEntity::class], version = 1)
-abstract class CartDatabase : RoomDatabase() {
-    abstract fun cartDao(): CartDao
-
-    companion object {
-        private const val DATABASE_NAME = "cart-db"
-
-        @Volatile
-        private var INSTANCE: CartDatabase? = null
-
-        fun getInstance(context: Context): CartDatabase {
-            return INSTANCE ?: synchronized(this) {
-                Room.databaseBuilder(
-                    context.applicationContext,
-                    CartDatabase::class.java,
-                    DATABASE_NAME,
-                ).build().also { INSTANCE = it }
-            }
-        }
-    }
-}
